pyproject.toml
[project]
name = "ocr-text-from-docs-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "njmin",email = "gazinurnurtdinov35@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.4,<4.0"
dependencies = [
    "aiogram (>=3.18.0,<4.0.0)",
    "emoji (>=2.14.1,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "arpakitlib (>=1.8.15,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
